requirements.txt
flask
google-generativeai
langchain
langchain-community
faiss-cpu
sentence-transformers
